refactor(mobile): rename metro-non-inlined-requires file
